:recycle: Remove DomainExceptions and update ManagedException class
Delete DomainException and update ManagedException for simple implementations
